feat: add user in repository
