+++ SocialNetworkApi/DataAccess/Repositories/Concretes/UserRepository.cs
@@ -1,0 +1,68 @@
+using Microsoft.EntityFrameworkCore;
+using SocialNetwork.Entities;
+using SocialNetwork.Persistence.DataBase;
+
+namespace SocialNetwork.Persistence.Repositories;
+
+public class UserRepository
+{
+    private readonly ApplicationDbContext _dbContext;
+
+    public UserRepository(ApplicationDbContext dbContext)
+    {
+        _dbContext = dbContext;
+    }
+
+    public async Task<User> CreateAsync(User user)
+    {
+        await _dbContext.Users.AddAsync(user);
+        await _dbContext.SaveChangesAsync();
+        return user;
+    }
+
+    public async Task<User?> GetByIdAsync(Guid userId)
+    {
+        return await _dbContext.Users.FindAsync(userId);
+    }
+
+    public async Task<User?> EditAsync(Guid userId, User updatedUser)
+    {
+        var existingUser = await _dbContext.Users.FindAsync(userId);
+        if (existingUser == null)
+            return null;
+
+        existingUser.Name = updatedUser.Name;
+        existingUser.Email = updatedUser.Email;
+        existingUser.Password = updatedUser.Password;
+        existingUser.ProfilePicture = updatedUser.ProfilePicture;
+
+        _dbContext.Users.Update(existingUser);
+        await _dbContext.SaveChangesAsync();
+        return existingUser;
+    }
+
+    public async Task DeleteAsync(Guid userId)
+    {
+        var user = await _dbContext.Users.FindAsync(userId);
+        if (user != null)
+        {
+            _dbContext.Users.Remove(user);
+            await _dbContext.SaveChangesAsync();
+        }
+    }
+
+    public async Task<IEnumerable<User>> GetAllAsync()
+    {
+        return await _dbContext.Users
+            .OrderBy(u => u.CreatedAt)
+            .ToListAsync();
+    }
+
+    public async Task<IEnumerable<User>> SearchAsync(string searchTerm)
+    {
+        return await _dbContext.Users
+            .Where(u => u.Name.Contains(searchTerm) || u.Email.Contains(searchTerm))
+            .OrderBy(u => u.Name)
+            .ToListAsync();
+    }
+}
